refactor,test: 테스트 패키지 변경, RacingCars 테스트 추가
