+++ src/test/java/step3/controller/RacingGameControllerTest.java
@@ -1,4 +1,4 @@
-package step3;
+package step3.controller;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
