+++ src/test/java/step3/domain/MoveStrategyTest.java
@@ -1,10 +1,7 @@
-package step3;
+package step3.domain;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
-import step3.domain.ForwardMoveStrategy;
-import step3.domain.MoveStrategy;
-import step3.domain.StopMoveStrategy;
 
 import static org.assertj.core.api.Assertions.assertThat;
